move scalar file
