+++ apps/hash-api/codegen.config.ts
@@ -1,5 +1,5 @@
 import { CodegenConfig } from "@graphql-codegen/cli";
-import { scalars } from "@local/hash-isomorphic-utils/scalar-mapping";
+import { scalars } from "@local/hash-isomorphic-utils/graphql/scalar-mapping";
 
 const config: CodegenConfig = {
   overwrite: true,
