+++ libs/@local/hash-isomorphic-utils/src/graphql/scalar-mapping.ts
